+++ src/event-listeners/CypressEventListener.ts
@@ -3,11 +3,17 @@ import { IEventListener } from './IEventListener'
 import location from '../utils/location'
 import gravityDocument from '../utils/gravityDocument'
 import IUserActionHandler from '../user-action/IUserActionHandler'
+import { ListenerFn } from 'eventemitter2'
 
 export default class CypressEventListener implements IEventListener {
-  private readonly listeners: Map<CypressEvent, any> = new Map()
+  private readonly listeners: Map<CypressEvent, ListenerFn> = new Map()
+  private readonly testAfterRunListener: (event: any) => void
 
   constructor(private readonly cypress: CypressObject, private readonly userActionHandler: IUserActionHandler) {
+    this.testAfterRunListener = () => {
+      console.log('[test:after:run] userActionHandler.flush()')
+      this.userActionHandler.flush()
+    }
     for (const cypressEvent of Object.values(CypressEvent)) {
       this.listeners.set(cypressEvent, (event: any) => {
         this.userActionHandler.handle({
@@ -26,9 +32,11 @@ export default class CypressEventListener implements IEventListener {
     for (const [cypressEvent, listener] of Array.from(this.listeners.entries())) {
       if (!this.cypress.listeners(cypressEvent).includes(listener)) this.cypress.addListener(cypressEvent, listener)
     }
+    this.cypress.addListener('test:after:run', this.testAfterRunListener)
   }
 
   terminate(): void {
+    this.cypress.removeListener('test:after:run', this.testAfterRunListener)
     for (const [cypressEvent, listener] of Array.from(this.listeners.entries())) {
       this.cypress.removeListener(cypressEvent, listener)
     }